Error if Person source includes Membership fields
When building Person + Membership information, we need to be very
careful not to include "Membership"-type data (party, constituency,
term, start/end date) from any Person source, as that will get merged
into every membership that the Person has had. This appears safe if
there's only a single term, and no-one changes party or area during
that, but the instant someone has a second membership, bad things start
to happen.

If we _do_ see any, when fetching the list of fields, raise an error.